% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{check_context}
\alias{check_context}
\title{Logical Check for Presence Context within Text}
\usage{
check_context(text, search, case_sensitive = FALSE)
}
\arguments{
\item{text}{A string value representing the text that is to be searched.}

\item{search}{The result of a \code{\link{construct_search}} call.}

\item{case_sensitive}{If TRUE, the function will ignore cases for all entities.}
}
\value{
A logical TRUE/FALSE value. If TRUE, the searched context is present within text, while FALSE indicates the absence of the defined context.
}
\description{
After constructing a search with \code{\link{construct_search}} you can use this function to do a logical test of whether or not the defined context is present within your text.
}
\examples{
text <- "The USA will intervene. A proposal has been submitted to the United Nations"
A <- c("USA", "Russia", "China")
B <- c("United Nations", "WTO", "NATO")
search <- construct_search(A, B, 50)
check_context(text, search)
}
\seealso{
\code{\link{construct_search}} for the construction of the search parameter.
}
