% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_context_all.R
\name{show_context_all}
\alias{show_context_all}
\title{Return a list that shows every context element found.}
\usage{
show_context_all(text, search, case_sensitive = FALSE)
}
\arguments{
\item{text}{A string value representing the text that is to be searched.}

\item{search}{The result of a \code{\link{construct_search}} call.}
}
\value{
The function will return a list of sub-strings of text which shows you where the contexts were found.
}
\description{
This function can be used to extract all text excerpts from a text, based on a context search constructed with \code{\link{check_context}}. It will show the entire sentence from the first occurance of the first relevant entity to the first sentence of the second entity.
}
\seealso{
\code{\link{construct_search}} for the construction of the search parameter.
}
